restructure sb_dl crate to move bigtable downloader
